LC-maximum-average-subarray-i [Runtime: 5 ms; Memory: 55.7 MB]

+++ array/maximum-average-subarray-i.java
@@ -1,12 +1,12 @@
 class Solution {
     public double findMaxAverage(int[] arr, int k) {
         int l=0,r=0;
-        float avg=0;
+        double avg=0;
         while(r<k) {
             avg=avg+arr[r];
             r++;
         }
-        float maxavg=avg / k;
+        double maxavg=avg / k;
         while(r<arr.length) {
             avg=avg+arr[r];
             avg=avg-arr[l];
